refactor: changing theme localization
