+++ src/pages/_app.tsx
@@ -3,7 +3,7 @@ import { DefaultSeo } from 'next-seo';
 
 // --- Chakra-UI ---
 import { ChakraProvider } from '@chakra-ui/react';
-import theme from '../../theme';
+import theme from '@styles/theme';
 
 // --- Analysis (optional) ---
 // import '@scripts/wdyr';
